backend/requirement.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
requests
langchain
boto3
pytest
